requirements.txt
py-clob-client
python-dotenv
web3
requests
